refactor: 인수테스트 패키지로 이동
