+++ src/test/java/nextstep/subway/path/acceptance/PathAcceptanceTest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.path;
+package nextstep.subway.path.acceptance;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
